requirements.txt
# For standalone audio recording
sounddevice
torch>=2.4.0
transformers>=4.53.0
gradio
pandas
pillow
opencv-python # Included as per your request
soundfile
librosa
python-dotenv
torchaudio